+++ src/views/authority/menu/index.vue
@@ -12,24 +12,24 @@
                 </el-col>
                 <el-col :xs="24" :sm="24" :md="24" :lg="17">
                     <div class="block-wrapper">
-                        <Block 
+                        <Block
                         ref="block"
                         :list-query-params.sync="listQueryParams"
-                        v-bind="blockAttrs" 
+                        v-bind="blockAttrs"
                         v-on="blockEvent"></Block>
                     </div>
                 </el-col>
            </el-row>
-           <!-- <Dialog 
-                v-bind="dialogAttrs" 
+           <!-- <Dialog
+                v-bind="dialogAttrs"
                 v-on:handleDialog="handleDialog"></Dialog> -->
-            <DialogForm 
+            <DialogForm
             ref="dialogform"
             v-if="innerVisible === true"
             v-bind="dialogAttrs"
             v-on:handleDialog="handleDialog"></DialogForm>
 
-            <DialogExcel 
+            <DialogExcel
             ref="dialogexcel"
             v-if="excelVisible === true"
             v-bind="excelAttrs"
@@ -111,7 +111,7 @@ export default {
             this.blockData = data.getLink()
             this.selectId = data.getID()
             if(this.selectId == '0'){
-                var a = this.menu.map(item=>{ return item.clone(item)})
+                var a = this.menu.map(item=>{ return item.clone(item)}).sort((x,y) => x.sort - y.sort)
                 this.blockData = a
             }
         },
@@ -128,22 +128,27 @@ export default {
                 { name:'取消',type:'info',status:'cancel'}]
                 this.innerVisible = true
                 this.dialogStatus = 'update'
-            }else if(index === 'delete'){
-                var isOk = await content.delete()
-                if(isOk){
+            }else if(index === 'delete'  || index === 'deleteMany'){
+                var isDelete = false
+                if(index === 'delete'){
+                  isDelete = await content.delete()
+                }else{
+                  var deleteArray = []
+                  content.forEach(item=>{
+                    deleteArray.push(item.id)
+                  })
+                  isDelete = await Menu.deleteMany(deleteArray.toString())
+                }
+                if(isDelete){
                     this.$message('刪除成功')
                     this.$store.dispatch('permission/setRoutes')
                     this.$socket.sendMsg('menus', 'routes' , '')
+                    this.$refs.block.clearSelectArray()
                 }else{
                     this.$message.error('系統錯誤')
                 }
             }else if(index === 'empty'){
-                // if(this.selectId == null){
-                //     this.$alert('如需新增子目錄，請先選擇父節點，再進行新增', '新增提醒', {
-                //         confirmButtonText: '確定'
-                //     })
-                // }
-                this.dialogData.push( Menu.empty() )    
+                this.dialogData.push( Menu.empty() )
                     this.dialogButtonsName = [
                     { name:'儲存',type:'primary',status:'create'},
                     { name:'取消',type:'info',status:'cancel'}]
@@ -168,21 +173,33 @@ export default {
                     this.excelVisible = true
                     this.excelType = 'uploadExcel'
                 }
+            }else if(index === 'updateMany'){
+              this.dialogStatus = 'updateMany'
+              content.forEach(item=>{
+                var obj = _.cloneDeep(item)
+                this.dialogData.push(obj)
+              })
+              this.dialogButtonsName = [
+                { name:'儲存',type:'primary',status:'updateManySave'},
+                { name:'取消',type:'info',status:'cancel'}]
+              this.innerVisible = true
             }
         },
         async handleDialog(title ,index, content){ //Dialog相關操作
             console.log(title ,index,JSON.stringify(content))
-            if(index == 'update' || index == 'create' || index === 'uploadExcelSave'){
+            if(index == 'update' || index == 'create' ||
+                index === 'uploadExcelSave' || index === 'updateManySave'){
                 delete content.linkMainMenus
                 delete content.linkAccessAuthorities
-                var isOk = index == 'update' ? await content.update() : 
-                index === 'create' ? await content.create(this.selectId == '0' ? null : this.selectId) : 
+                var isOk = index == 'update' || index == 'updateManySave' ? await content.update() :
+                index === 'create' ? await content.create(this.selectId == '0' ? null : this.selectId) :
                 await Menu.postMany(this.selectId,content)
                 if(isOk){
-                    index == 'update' ? this.$message('更新成功') : this.$message('新增成功')
+                    index == 'update' || index == 'updateManySave' ?
+                        this.$message('更新成功') : this.$message('新增成功')
                     this.$store.dispatch('permission/setRoutes')
                     this.$socket.sendMsg('menus', 'routes' , '')
-                    this.innerVisible = false
+                    if(index !== 'updateManySave') this.innerVisible = false
                     this.excelVisible = false
                 }else{
                     this.$message.error('系統錯誤')
@@ -190,6 +207,7 @@ export default {
             }else{
                 this.innerVisible = false
                 this.excelVisible = false
+                this.$refs.block.clearSelectArray()
             }
         },
         async changeTable(value){
