+++ src/views/authority/roles/index.vue
@@ -2,33 +2,33 @@
         <div class="editor-container">
             <el-col :xs="24" :sm="24" :md="24" :lg="24">
                 <div class="block-wrapper">
-                    <Block 
+                    <Block
                         ref="block"
                         :list-query-params.sync="listQueryParams"
-                        v-bind="blockAttrs" 
+                        v-bind="blockAttrs"
                         v-on="blockEvent"></Block>
                 </div>
             </el-col>
-            <!-- <Dialog 
+            <!-- <Dialog
                 ref="dialog"
-                v-bind="dialogAttrs" 
+                v-bind="dialogAttrs"
                 :accessAuthoritiesData="treeData"
                 :accessAuthorities="roleAccessAuthority"
                 v-on:handleDialog="handleDialog"></Dialog> -->
-            
-            <DialogForm 
+
+            <DialogForm
             ref="dialogform"
             v-if="innerVisible === true"
             v-bind="dialogAttrs"
             v-on:handleDialog="handleDialog"></DialogForm>
 
-            <DialogExcel 
+            <DialogExcel
             ref="dialogexcel"
             v-if="excelVisible === true"
             v-bind="excelAttrs"
             v-on:handleDialog="handleDialog"></DialogExcel>
 
-            <DialogAuthority 
+            <DialogAuthority
             ref="dialogauthority"
             v-if="authorityVisible === true"
             v-bind="authorityAttrs"
@@ -105,7 +105,7 @@ export default {
             await this.getAllRole()
         },
         async getAllRole(){
-            this.listQueryParams.orderBy = 'sort'    
+            this.listQueryParams.orderBy = 'sort'
             var data = await Role.getSearchPage(this.listQueryParams)
             this.blockData = data.result
             this.listQueryParams.total = data.totalPageCount
@@ -143,18 +143,27 @@ export default {
                 { name:'取消',type:'info',status:'cancel'}]
                 this.innerVisible = true
                 this.dialogStatus = 'update'
-            }else if(index === 'delete'){
-                var isOk = await content.delete()
-                if(isOk){
+            }else if(index === 'delete' || index === 'deleteMany'){
+                var isDelete = false
+                if(index === 'delete'){
+                  isDelete = await content.delete()
+                }else{
+                  var deleteArray = []
+                  content.forEach(item=>{
+                    deleteArray.push(item.id)
+                  })
+                  isDelete = await Role.deleteMany(deleteArray.toString())
+                }
+                if(isDelete){
                     this.$message('刪除成功')
-                    // this.$store.dispatch('permission/setmenu',await  Menu.get())
-                    // this.$store.dispatch('building/setroles')
                     if(this.listQueryParams.pageIndex !== 1 && this.blockData.length == 1){
                         this.listQueryParams.pageIndex = this.listQueryParams.pageIndex-1
                     }
                     this.$store.dispatch('building/setroles')
-                    this.$socket.sendMsg('roles','delete',content.getID())
+                    this.$socket.sendMsg('roles','delete' ,
+                        index === 'delete'  ? content.getID() : deleteArray.toString())
                     await this.getAllRole()
+                    this.$refs.block.clearSelectArray()
                 }else{
                     this.$message.error('系統錯誤')
                 }
@@ -178,12 +187,23 @@ export default {
             }else if(index === 'uploadExcel'){
                 this.excelVisible = true
                 this.excelType = 'uploadExcel'
+            }else if(index === 'updateMany'){
+              this.dialogStatus = 'updateMany'
+              content.forEach(item=>{
+                var obj = _.cloneDeep(item)
+                this.dialogData.push(obj)
+              })
+              this.dialogButtonsName = [
+                { name:'儲存',type:'primary',status:'updateManySave'},
+                { name:'取消',type:'info',status:'cancel'}]
+              this.innerVisible = true
             }
         },
         async handleDialog(title ,index, content){ //Dialog相關操作
             console.log(title , index, JSON.stringify(content))
-            if(index === 'update' || index === 'create' || index === 'uploadExcelSave'){
-                var result = index === 'update' ? await content.update() : 
+            if(index === 'update' || index === 'create' ||
+            index === 'uploadExcelSave' || index === 'updateManySave'){
+                var result = index === 'update' || index === 'updateManySave' ? await content.update() :
                 index === 'create' ? await content.create() : await Role.postMany(content)
                 var condition = index !== 'uploadExcelSave' ? Object.keys(result).length !== 0 : result.result.length !== 0
                 if(condition){
@@ -191,7 +211,7 @@ export default {
                     this.$store.dispatch('building/setroles')
                     this.$socket.sendMsg('roles', index , index !== 'uploadExcelSave' ? result: result.result)
                     await this.getAllRole()
-                    this.innerVisible = false
+                    if(index !== 'updateManySave') this.innerVisible = false
                     this.excelVisible = false
                 }else{
                     if(index !== 'uploadExcelSave'){
@@ -219,7 +239,7 @@ export default {
                             count++
                         }
                     }
-                    if (count === 0) {   
+                    if (count === 0) {
                         remove.push(stra)
                     }
                 })
@@ -232,7 +252,7 @@ export default {
                             count++
                         }
                     }
-                    if (count === 0) {   
+                    if (count === 0) {
                         add.push(stra)
                     }
                 })
@@ -276,8 +296,8 @@ export default {
                 this.innerVisible = false
                 this.excelVisible = false
                 this.authorityVisible = false
+                this.$refs.block.clearSelectArray()
             }
-            
         },
         async changeTable(value){
             this.isTable = value
